Moved code from CaseManagementViewController to new controllers and added CaseFileViewController
